Moving authentication logic to separate service class.
