+++ Application/Services/IAuthenticationService.cs
@@ -1,0 +1,8 @@
+using Domain.DTO.Response;
+
+namespace Application.Services;
+
+public interface IAuthenticationService
+{
+    Task<TokenDTO> Authenticate(AuthenticationDTO authenticationRequestBody);
+}
