+++ Application/Services/IServiceManager.cs
@@ -6,5 +6,6 @@ public interface IServiceManager
 {
     IBaseService BaseService { get; }
     IQuestionService QuestionService { get; }
+    IAuthenticationService AuthenticationService{ get; }
     IMapper Mapper { get; }
 }
