+++ Domain/DTO/Request/AuthenticationDTO.cs
@@ -1,0 +1,11 @@
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+public class AuthenticationDTO
+{
+    public string? Email { get; set; }
+    public string? Password { get; set; }
+}
