+++ Domain/DTO/Response/QuestionDetailedDTO.cs
@@ -1,4 +1,4 @@
-namespace Domain.DTO;
+namespace Domain.DTO.Response;
 
 public class QuestionDetailedDTO : QuestionDTO
 {
