+++ Domain/DTO/Response/TokenDTO.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Domain.DTO
+namespace Domain.DTO.Response
 {
     public class TokenDTO
     {
